Refactored the components into seperate subdirs for better organization.
